# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  CreateElectricMeterRequest:
    properties:
      billing_month:
        type: string
      billing_type:
        $ref: '#/definitions/ElectricBillingType'
      current_reading:
        type: number
      flat_amount:
        type: number
      note:
        type: string
      previous_reading:
        type: number
      reading_date:
        type: string
      room_id:
        type: string
      unit_price:
        type: number
    type: object
  CreateExpenseRequest:
    properties:
      amount:
        type: number
      category:
        $ref: '#/definitions/ExpenseCategory'
      description:
        type: string
      expense_date:
        type: string
      note:
        type: string
    type: object
  CreateMaintenanceRequestRequest:
    properties:
      description:
        type: string
      note:
        type: string
      priority:
        $ref: '#/definitions/MaintenancePriority'
      reported_date:
        type: string
      resolved_date:
        type: string
      room_id:
        type: string
      status:
        $ref: '#/definitions/MaintenanceStatus'
      title:
        type: string
    type: object
...
  ElectricBillingType:
    enum:
    - meter
    - flat
    type: string
    x-enum-varnames:
    - ElectricBillingTypeMeter
    - ElectricBillingTypeFlat
  ElectricMeter:
    properties:
      billing_month:
        type: string
      billing_type:
        $ref: '#/definitions/ElectricBillingType'
      created_at:
        type: string
      created_by:
        type: string
      current_reading:
        type: number
      flat_amount:
        type: number
      id:
        type: string
      note:
        type: string
      previous_reading:
        type: number
      reading_date:
        type: string
      room_id:
        type: string
      unit_price:
        type: number
      updated_at:
        type: string
      updated_by:
        type: string
      updated_by_name:
        type: string
    type: object
  ElectricMeterDetail:
    properties:
      billing_month:
        type: string
      billing_type:
        $ref: '#/definitions/ElectricBillingType'
      created_at:
        type: string
      created_by:
        type: string
      current_reading:
        type: number
      flat_amount:
        type: number
      id:
        type: string
      note:
        type: string
      previous_reading:
        type: number
      reading_date:
        type: string
      room_id:
        type: string
      room_number:
        type: string
      total_amount:
        type: number
      unit_price:
        type: number
      unit_used:
        type: number
      updated_at:
        type: string
      updated_by:
        type: string
      updated_by_name:
        type: string
    type: object
  Expense:
    properties:
      amount:
        type: number
      category:
        $ref: '#/definitions/ExpenseCategory'
      created_at:
        type: string
      description:
        type: string
      dormitory_id:
        type: string
      expense_date:
        type: string
      id:
        type: string
      note:
        type: string
      updated_at:
        type: string
    type: object
  ExpenseCategory:
    enum:
    - repair
    - utilities
    - supplies
    - salary
    - other
    type: string
    x-enum-varnames:
    - ExpenseCategoryRepair
    - ExpenseCategoryUtilities
    - ExpenseCategorySupplies
    - ExpenseCategorySalary
    - ExpenseCategoryOther
...
  MaintenancePriority:
    enum:
    - low
    - normal
    - high
    - urgent
    type: string
    x-enum-varnames:
    - MaintenancePriorityLow
    - MaintenancePriorityNormal
    - MaintenancePriorityHigh
    - MaintenancePriorityUrgent
  MaintenanceRequest:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      note:
        type: string
      priority:
        $ref: '#/definitions/MaintenancePriority'
      reported_date:
        type: string
      resolved_date:
        type: string
      room_id:
        type: string
      status:
        $ref: '#/definitions/MaintenanceStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  MaintenanceRequestDetail:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      note:
        type: string
      priority:
        $ref: '#/definitions/MaintenancePriority'
      reported_date:
        type: string
      resolved_date:
        type: string
      room_id:
        type: string
      room_number:
        type: string
      status:
        $ref: '#/definitions/MaintenanceStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  MaintenanceStatus:
    enum:
    - open
    - in_progress
    - done
    - cancelled
    type: string
    x-enum-varnames:
    - MaintenanceStatusOpen
    - MaintenanceStatusInProgress
    - MaintenanceStatusDone
    - MaintenanceStatusCancelled
...
  UpdateElectricMeterRequest:
    properties:
      billing_month:
        type: string
      billing_type:
        $ref: '#/definitions/ElectricBillingType'
      current_reading:
        type: number
      flat_amount:
        type: number
      note:
        type: string
      previous_reading:
        type: number
      reading_date:
        type: string
      unit_price:
        type: number
    type: object
  UpdateExpenseRequest:
    properties:
      amount:
        type: number
      category:
        $ref: '#/definitions/ExpenseCategory'
      description:
        type: string
      expense_date:
        type: string
      note:
        type: string
    type: object
  UpdateMaintenanceRequestRequest:
    properties:
      description:
        type: string
      note:
        type: string
      priority:
        $ref: '#/definitions/MaintenancePriority'
      reported_date:
        type: string
      resolved_date:
        type: string
      room_id:
        type: string
      status:
        $ref: '#/definitions/MaintenanceStatus'
      title:
        type: string
    type: object
...
info:
  contact: {}
  title: API
paths:
...
  /electric-meters:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ElectricMeter'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายการค่ามิเตอร์ไฟฟ้า
      tags:
      - electric-meters
    post:
      consumes:
      - application/json
      parameters:
      - description: Electric meter payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateElectricMeterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ElectricMeter'
      security:
      - ApiKeyAuth: []
      summary: บันทึกค่ามิเตอร์ไฟฟ้า
      tags:
      - electric-meters
  /electric-meters/{id}:
    delete:
      parameters:
      - description: Electric Meter Reading ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบค่ามิเตอร์ไฟฟ้า
      tags:
      - electric-meters
    get:
      parameters:
      - description: Electric Meter Reading ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ElectricMeterDetail'
      security:
      - ApiKeyAuth: []
      summary: ดูค่ามิเตอร์ไฟฟ้าตาม ID
      tags:
      - electric-meters
    put:
      consumes:
      - application/json
      parameters:
      - description: Electric Meter Reading ID
        in: path
        name: id
        required: true
        type: string
      - description: Electric meter payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateElectricMeterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ElectricMeter'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขค่ามิเตอร์ไฟฟ้า
      tags:
      - electric-meters
  /electric-meters/latest:
    get:
      parameters:
      - description: Room ID
        in: query
        name: room_id
        required: true
        type: string
      - description: Billing month (YYYY-MM-DD)
        in: query
        name: billing_month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ElectricMeterDetail'
      security:
      - ApiKeyAuth: []
      summary: ดูค่ามิเตอร์ไฟฟ้าล่าสุดของห้อง
      tags:
      - electric-meters
  /expenses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Expense'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายการค่าใช้จ่าย
      tags:
      - expenses
    post:
      consumes:
      - application/json
      parameters:
      - description: Expense payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateExpenseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Expense'
      security:
      - ApiKeyAuth: []
      summary: สร้างค่าใช้จ่าย
      tags:
      - expenses
  /expenses/{id}:
    delete:
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบค่าใช้จ่าย
      tags:
      - expenses
    get:
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Expense'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลค่าใช้จ่ายตาม ID
      tags:
      - expenses
    put:
      consumes:
      - application/json
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      - description: Expense payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Expense'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขค่าใช้จ่าย
      tags:
      - expenses
  /maintenance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/MaintenanceRequest'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายการแจ้งซ่อม
      tags:
      - maintenance
    post:
      consumes:
      - application/json
      parameters:
      - description: Maintenance request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateMaintenanceRequestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/MaintenanceRequest'
      security:
      - ApiKeyAuth: []
      summary: สร้างรายการแจ้งซ่อม
      tags:
      - maintenance
  /maintenance/{id}:
    delete:
      parameters:
      - description: Maintenance Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบรายการแจ้งซ่อม
      tags:
      - maintenance
    get:
      parameters:
      - description: Maintenance Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MaintenanceRequestDetail'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลแจ้งซ่อมตาม ID
      tags:
      - maintenance
    put:
      consumes:
      - application/json
      parameters:
      - description: Maintenance Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Maintenance request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateMaintenanceRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MaintenanceRequest'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขรายการแจ้งซ่อม
      tags:
      - maintenance
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"